Create books_list.py
I was looking for an API that gives me the list of all books in the protestant bible. I can improve this by allowing a user to specify the bible they want and it gives them a list of books in that bible. 
The books are ordered chronologically too. I can improve this by making the books a key-pair value that states the sequential order in the respective bibles.